Rename tex_common.h to nmgltex_common.h
Rename include file to be consistent with common style
