+++ include/nmgltex_common.h
